Fix build warnings across KuMenuSet.cpp, NTFSLink.cpp, and pugxml.h
**KuMenuSet.cpp**

-   Added `#define PUGAPI_VARIANT`, `PUGAPI_VERSION_MAJOR`, and `PUGAPI_VERSION_MINOR` before `#include "pugxml.h"` to silence the "API variant/major version mismatch" compiler messages.
-   Fixed C4018 (signed/unsigned mismatch): changed `int i` to `size_t i` in the `node.attributes()` loop, and added `(int)` cast when comparing `int iFrom` against `size_t count` in three `for` loops.
-   Fixed C4267 (size\_t → unsigned int): added `(unsigned int)i` casts on the three `node.attribute(i)` call sites, which became necessary after the loop variable was widened to `size_t`.
-   Fixed C4267: added `(int)` cast on `aCmds.GetCount()` assigned to an `int`.

**NTFSLink.cpp**

-   Fixed C4267 (size\_t → DWORD): added `(DWORD)` cast on the `wcslen(...)` result assigned to `len`.
-   Fixed C4244 (DWORD/size\_t → USHORT): added `(USHORT)` casts on the three `REPARSE_DATA_BUFFER` field assignments — `SubstituteNameLength`, `PrintNameOffset`, and `ReparseDataLength`.

**pugxml.h**

-   Fixed C4267 (size\_t → unsigned int): added `(unsigned int)` casts on the two `min(valuelen, ...)` expressions assigned to `const unsigned int n` (x64 build only).

+++ src/pugxml.h
@@ -2477,9 +2477,9 @@ public:
 				{
 					const unsigned int n =
 #ifdef PUGOPT_NONSEG
-						min(valuelen,node->value_size);
+						(unsigned int)min(valuelen,node->value_size);
 #else
-						min(valuelen,_tcslen(node->value));
+						(unsigned int)min(valuelen,_tcslen(node->value));
 #endif
 					_tcsncpy(value,node->value,n);
 					value[n] = 0;
